+++ src/utils.js
@@ -23,28 +23,41 @@ function addObjectToArray(arr, obj, unshift = false, avoidDuplicates = true) {
  */
 function prettyDate(dateString) {
   const date = new Date(dateString)
-  return new Intl.DateTimeFormat('default').format(date)
+  return new Intl.DateTimeFormat(navigator.language, {dateStyle: 'short'}).format(date)
+}
+
+/**
+ * input: '2023-12-27T17:08:19.021410+01:00'
+ * output: '12/27/2023, 8:19AM'
+ */
+function prettyDateTime(dateTimeString) {
+  const date = new Date(dateTimeString)
+  return new Intl.DateTimeFormat(navigator.language, {dateStyle: 'short', timeStyle: 'short'}).format(date)
 }
 
 /**
  * https://johnresig.com/blog/javascript-pretty-date/
  * input: '2023-12-27T17:08:19.021410+01:00'
  * output: '5 hours ago', 'Yesterday', '2 days ago'...
+ * changes: add short (replace 'days' with 'd', remove 'Yesterday') & shortest (remove 'ago'), extend days & weeks
  */
-function prettyRelativeDateTime(dateTimeString, short=false) {
+function prettyRelativeDateTime(dateTimeString, size=null) {
   var date = new Date((dateTimeString || "").replace(/-/g, "/").replace(/[TZ]/g, " ")),
       diff = (((new Date()).getTime() - date.getTime()) / 1000),
       day_diff = Math.floor(diff / 86400);
 
-  if (isNaN(day_diff) || day_diff < 0 || day_diff >= 31) return;
+  if (isNaN(day_diff) || day_diff < 0) return;
 
-  if (short) {
-    // replace 'days' with 'd', remove 'Yesterday'
+  if (size == 'shortest') {
     return day_diff == 0 && (
-      diff < 60 && "just now" || diff < 120 && "1m ago" || diff < 3600 && Math.floor(diff / 60) + "m ago" || diff < 7200 && "1h ago" || diff < 86400 && Math.floor(diff / 3600) + "h ago") || day_diff < 7 && day_diff + "d ago" || day_diff < 31 && Math.ceil(day_diff / 7) + "w ago";
+      diff < 60 && "just now" || diff < 120 && "1m" || diff < 3600 && Math.floor(diff / 60) + "m" || diff < 7200 && "1h" || diff < 86400 && Math.floor(diff / 3600) + "h") || day_diff < 10 && day_diff + "d" || Math.ceil(day_diff / 7) + "w";
+  }
+  if (size == 'short') {
+    return day_diff == 0 && (
+      diff < 60 && "just now" || diff < 120 && "1m ago" || diff < 3600 && Math.floor(diff / 60) + "m ago" || diff < 7200 && "1h ago" || diff < 86400 && Math.floor(diff / 3600) + "h ago") || day_diff < 10 && day_diff + "d ago" || Math.ceil(day_diff / 7) + "w ago";
   }
   return day_diff == 0 && (
-  diff < 60 && "just now" || diff < 120 && "1 minute ago" || diff < 3600 && Math.floor(diff / 60) + " minutes ago" || diff < 7200 && "1 hour ago" || diff < 86400 && Math.floor(diff / 3600) + " hours ago") || day_diff == 1 && "Yesterday" || day_diff < 7 && day_diff + " days ago" || day_diff < 31 && Math.ceil(day_diff / 7) + " weeks ago";
+  diff < 60 && "just now" || diff < 120 && "1 minute ago" || diff < 3600 && Math.floor(diff / 60) + " minutes ago" || diff < 7200 && "1 hour ago" || diff < 86400 && Math.floor(diff / 3600) + " hours ago") || day_diff == 1 && "Yesterday" || day_diff < 10 && day_diff + " days ago" || Math.ceil(day_diff / 7) + " weeks ago";
 }
 
 function getCategory(categoryId) {
@@ -68,6 +81,7 @@ function getLocationTitle(locationObject, withEmoji=false) {
 export default {
   addObjectToArray,
   prettyDate,
+  prettyDateTime,
   prettyRelativeDateTime,
   getCategory,
   getCountryEmojiFromName,