+++ src/views/ProductDetail.vue
@@ -1,18 +1,26 @@
 <template>
   <v-row>
     <v-col cols="12" sm="6">
-      <PriceCard v-if="product" :product="product" :readonly="true" elevation="1"></PriceCard>
-      <v-card v-if="productIsCategory" :title="getCategoryName" prepend-icon="mdi-fruit-watermelon"></v-card>
+      <PriceCard v-if="!loading && !productIsCategory" :product="product" :readonly="true" elevation="1"></PriceCard>
+      <v-card v-if="!loading && productIsCategory" :title="getCategoryName" prepend-icon="mdi-fruit-watermelon" elevation="1"></v-card>
     </v-col>
   </v-row>
 
-  <v-row class="mt-0" v-if="product">
+  <v-row class="mt-0" v-if="!productNotFound">
     <v-col cols="12" sm="6">
       <v-btn v-if="product.code && product.source" size="small" append-icon="mdi-open-in-new" :href="getProductOFFUrl(product)" target="_blank">
         Open Food Facts
       </v-btn>
-      <p v-if="!product.code || !product.source" class="text-red">
+    </v-col>
+  </v-row>
+
+  <v-row class="mt-0" v-if="productOrCategoryNotFound">
+    <v-col cols="12" sm="6">
+      <p v-if="productNotFound" class="text-red">
         <i>Product not found in Open Food Facts... Don't hesitate to add it :)</i>
+      </p>
+      <p v-if="categoryNotFound" class="text-red">
+        <i>Category not found...</i>
       </p>
     </v-col>
   </v-row>
@@ -49,8 +57,8 @@ export default {
   },
   data() {
     return {
-      productId: this.$route.params.id,
-      product: null,
+      productId: this.$route.params.id,  // product_code or product_category
+      product: { code: this.$route.params.id },
       productPriceList: [],
       productPriceTotal: null,
       productPricePage: 0,
@@ -66,16 +74,23 @@ export default {
       return this.productId.startsWith('en')
     },
     getCategoryName() {
-      if (this.productIsCategory) {
-        const tag = utils.getCategory(this.productId)
-        return tag ? tag.name : this.productId
-      }
+      const tag = this.getCategory()
+      return tag ? tag.name : this.productId
     },
+    productNotFound() {
+      return !this.productIsCategory && (!this.product.code || !this.product.source)
+    },
+    categoryNotFound() {
+      return this.productIsCategory && !!this.getCategory
+    },
+    productOrCategoryNotFound() {
+      return this.productNotFound || this.categoryNotFound
+    }
   },
   methods: {
     getProduct() {
       if (!this.productIsCategory) {
-        return api.getProductById(this.productId)
+        return api.getProductByCode(this.productId)
           .then((data) => {
             if (data.id) {
               this.product = data
@@ -83,10 +98,15 @@ export default {
           })
       }
     },
+    getCategory() {
+      if (this.productIsCategory) {
+        return utils.getCategory(this.productId)
+      }
+    },
     getProductPrices() {
       this.loading = true
       this.productPricePage += 1
-      return api.getPrices({ [this.productIsCategory ? 'category_tag' : 'product_id']: this.productId, page: this.productPricePage })
+      return api.getPrices({ [this.productIsCategory ? 'category_tag' : 'product_code']: this.productId, page: this.productPricePage })
         .then((data) => {
           this.productPriceList.push(...data.items)
           this.productPriceTotal = data.total